aft: NHG-ID and NH-ID space management
In some systems NHG & NH ID generation is internal so it would
result in mismatch between index pushed by the client and index
created by the system.

Adding a new leaf in the next-hop-group & next-hop AFT entry to
represent the client published value. This helps the client to
create an association between configured & extracted index.