Add ip-address-zoned union typedef for link-local neighbor addresses
Similar to the existing typedef for ip-address, this new typedef
provides a union between existing ipv4-address-zoned and
ipv6-address-zoned types that are used by link-local peering.

Currently systems such as Arista's EOS support link-local peering.
In particular when dealing with BGP IPv6 link-local peering, the current
neighbor-address key is incompatible as it requires an ip-address type.
With this new union typedef, it may be possible to update the
neighbor-address key for BGP peer information to use this new
"ip-address-zoned" key instead to allow supporting link-local peers.
This change should be backwards compatible as it simply relaxes the
restrictions on the key type.