Create terminal-device-properties-guide.md
This document includes a detail explanation of the device-manifest concept and in particular on how it should be used the proposed openconfig-terminal-device-properties.yang model.

This proposal was submitted by the Telecom Infra Project (TIP) Mandatory Use Cases for SDN Transport (MUST) sub-group. This is an operator centric initiative which intends to achieve a wider consensus about the SDN standards to use on the network transport segment.