Add fib-ready container under afts container
AFT streaming over gNMI is an eventually consistent system.
When the device updates an entry it is usually expected to
stream an update to the client within a vert short amount
of time (few milliseconds). So, Telemetry collector or a
controller that parse the AFT doesn't have a consistent
snapshot, or overall versioned copy of AFT with the device
at any point of time.

In certain failure modes like device boot up, gNMI daemon
failure and device/routing engine stateful switchover
Telemetry collector or a controller need a flag to
determine whether it is in consistent with the device or
not such that it can a corrective action when needed.
A device would set this leaf or flag to indicate to the
client that AFT data/view is consistent.