* (M) release/models/system/system-aaa.yang * (M) release/models/system/system-aaa-radius.yang * (M) release/models/system/system-aaa-tacacs.yang
 - Update role to be mandatory attribute of user configuration

This is to address the undefined behavior when a system user is created but no role is specified.

1. Junos/EVO require a role and do not allow a user to be created without one.
2. EOS allows the creation of a user without specifying a role.  The default role is 'network-operator',
with a privilege level of 1. (https://www.arista.com/en/um-eos/eos-user-security#xx1347271)
3. Can't find documentation for IOS-XR, but NX-OS behaves in a similiar manner to EOS and assigns a
default role of network-operator to a user if a role isn't provided at time of creation (role of 'operator').