Fix `.spec.yml` where name doesn't match module name to fix online documentation.
The way the path links at
https://openconfig.net/projects/models/paths/index.html works is that
the root module of each path is assumed to be the name of the HTML file
containing the documentation of the path.

For example, the root module of `/interfaces/interface/state/name` is
`openconfig-interfaces`, so the link refers to
`openconfig-interfaces.html`.

This means that for this to work, there must be
an entry in *some* `spec.yml` file whose name matches the root module
for the path.

For example, the above path points to a valid link because there is a
spec file with the following:
```yaml
- name: openconfig-interfaces
  docs:
      - yang/interfaces/openconfig-interfaces.yang
```

This means that if we changed the `name` field in the above YAML to
`openconfig-intf`, then the path no longer points to a valid link.

Ideally, for each path we know the YAML entry which compiled it, but
this is hard to do since we would have to analyze every YAML spec
entry.

**Summary:** This PR fixes up the `.spec.yml` files such that the above
informal rule is used so that almost all paths now point to valid links.