Change NTP offset, root-delay, and root-dispersion to int64 nanoseconds (#1022)
As described in RFC 5905, the range of the offset that can be computed
from the timestamps in NTP packets is +/- 34 years.  A value of int64
nanoseconds has a range of +/- 2^63 seconds / 10^9, which is
+/- 292 years -- more than enough.

Root-delay and root-dispersion values in NTP packets have a smaller
range -- about 18 hours -- so the same type and units are also
sufficient for them.

Also update the descriptions for accuracy and clarity:

* The old description of root-delay matched the (single-server =
  non-"root") delay, but the *root* delay is accumulated across strata.
* The old description of root-dispersion was not clearly differentiated
  from the (single-server = non-"root") dispersion.

This change is not backwards compatible, so the major version is
incremented.