gnsi: align with latest certz version
The certz yang file was updated since 
the PR to move from openconfig/gnsi to 
openconfig/public, and this updated
version of certz had not been
included in the PR.

see the changes to gnsi-certz.yang at:
https://github.com/openconfig/gnsi/pull/156

This commit also sets the container
name for the certz counters to be just
"counters" to be backward-compatible 
with the existing gnsi-certz.yang.

The module name/prefix already 
mark these counters as certz-related,
so we don't need to duplicate this
information in the container name too.