Change of key's leafref references in "compatible-transceiver" and "transceiver-compatible-modes" lists
Keys of "compatible-transceiver" and "transceiver-compatible-modes" lists cannot reference an existing key leaf (a leaf which is being used as a key in another list) such as the case of the "mode-id" in "operational-mode-descriptors" list and the "component-descriptor-id" in "transceiver-descriptors" list.
Thus, the leafrefs have been substituted by String or UINT16 leafs to accomodate an absolute reference attribute.
Moreover, the keys have been modeled as leafref of the state leafs, following the tranditional openconfig style for lists modelling.