Add status flags for transceiver host lanes and media channels (#1197)
These flags are defined in the CMIS spec.  Some of them are also defined
in the earlier specs for management interfaces of SFP+ and QSFP+
modules.

Some of the flags are specific to media channels.  They are defined in
the existing list /components/component/transceiver/physical-channels/channel.

Other flags are specific to electrical host lanes.  This change adds a
new analogous list /components/component/transceiver/host-lanes/lane to
hold them.